Add a rotating newsticker and use it instead of the scrolling one
The rotating newsticker consumes less CPU, since only moves for 1 second
each 5 seconds instead of scrolling continuously.

The drawback is that javascript is used to change the active item each
second, so it uses a javscript timer, while the scrolling newsticker
only used javascript for the setup and then used only css effects.